tests pass

+++ routes/folder.js
@@ -5,13 +5,14 @@ const express = require('express');
 const router = express.Router();
 const mongoose = require('mongoose');
 const Folder = require('../models/folder');
+let Note = require('../models/note');
 mongoose.Promise = global.Promise;
 
 
 
 
 
-router.get('/folder',(req,res,next) => {
+router.get('/',(req,res,next) => {
   let searchTerm;
   let filterArr = [];
   if (searchTerm) {
@@ -29,20 +30,32 @@ router.get('/folder',(req,res,next) => {
 });
 
 
-router.get('/folder/:id', (req, res, next) => {
+router.get('/:id', (req, res, next) => {
+  
   let {id} = req.params;
-       
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    const err = new Error('Invalid \':id\'');
+    err.status = 400;
+    return next(err);
+  }
+  
   Folder.findById(id)
     .then(results => {
-      res.json(results);
+      if(results){
+        res.json(results);}
     }).catch(err => {
       next(err);
     });
 });
 
 
-router.post('/folder', (req, res, next) => {
+router.post('/', (req, res, next) => {
   let {name} = req.body;
+  if (!name){{
+    const err = new Error('Missing `name` in request body');
+    err.status = 400;
+    return next(err);
+  }}
   let newFolder={
     name
   };
@@ -50,15 +63,31 @@ router.post('/folder', (req, res, next) => {
     .then((results) => {
       res.location(`${req.originalUrl}/${res.id}`).status(201).json(results);
     })
-    .catch(err=>next(err));
+    .catch(err=>  {
+      if (err.code === 11000) {
+        err = new Error('The folder name already exists');
+        err.status = 400;
+      }
+      next(err);
+    });
 });
 
-router.put('/folder/:id', (req, res, next) => {
+router.put('/:id', (req, res, next) => {
   let {id} = req.params;
-    
-    
-  let upObj = {};
-  if(req.body.name){upObj.title = req.body.name;}
+  let {name} = req.body;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    const err = new Error('Invalid \':id\'');
+    err.status = 400;
+    return next(err);
+  }
+  /***** Never trust users - validate input *****/
+  if (!name) {
+    const err = new Error('Missing `name` in request body');
+    err.status = 400;
+    return next(err);
+  }
+  let upObj = {name};
+  if(req.body.name){upObj.name = req.body.name;}
   
   return Folder.findByIdAndUpdate(id,{$set:upObj},{upsert:true, new:true})
     .then(results=>{
@@ -66,13 +95,23 @@ router.put('/folder/:id', (req, res, next) => {
     }).catch(err => next(err));
 });
 
-router.delete('/folder/:id', (req, res) => {
+router.delete('/:id', (req, res,next) => {
   let {id} = req.params;
-  Folder.deleteOne({_id:id})
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    const err = new Error('Invalid \':id\'');
+    err.status = 400;
+    return next(err);
+  }
+  Note.updateMany({folderId : id},{$unset: {folderId: null}})
     .then(()=>{
-      res.status(204).end();
+      return Folder.deleteOne({_id:id});
+    }).then(result => {
+      if(result){res.status(204).end();}
+      else{next();}
     });
+      
 });
+
